Refactor BaseTableCRUDTest to split the testing table preparation into a new base class BaseTableCRUDTestEnv. This new base class BaseTableCRUDTestEnv will also be used by a new test for table commits.
