+++ server/src/test/java/io/unitycatalog/server/sdk/tables/SdkTableCRUDTest.java
@@ -73,7 +73,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
 
   @Test
   public void testListTablesWithNoNextPageTokenShouldReturnNull() throws Exception {
-    createCommonResources();
     TableInfo testingTable =
         createTestingTable(TestUtils.TABLE_NAME, TestUtils.STORAGE_LOCATION, tableOperations);
     ListTablesResponse resp =
@@ -90,7 +89,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
 
   @Test
   public void testListTablesWithNextPageTokenShouldReturnNextPageToken() throws Exception {
-    createCommonResources();
     List<TableInfo> testingTables = createMultipleTestingTables(11);
     ListTablesResponse resp =
         localTablesApi.listTables(
@@ -114,7 +112,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
    */
   @Test
   public void testStagingTableCreationAndManagedTableFromStagingLocation() throws Exception {
-    createCommonResources();
     String stagingTableName = "staging_test_table";
 
     // Step 1: Create a staging table
@@ -200,8 +197,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
    */
   @Test
   public void testManagedTableCreationFromNonExistentStagingLocationShouldFail() throws Exception {
-    createCommonResources();
-
     // Use a fake staging location that doesn't exist
     String fakeLocationUuid = "00000000-0000-0000-0000-000000000000";
     String fakeStagingLocation = "file:///tmp/ucroot/tables/" + fakeLocationUuid;
@@ -228,8 +223,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
   /** Test that attempting to create a duplicate staging table fails with ALREADY_EXISTS error. */
   @Test
   public void testDuplicateStagingTableCreationShouldFail() throws Exception {
-    createCommonResources();
-
     // Create an external table
     String externalTableName = "duplicate_table";
     CreateTable createTableRequest =
@@ -286,8 +279,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
    */
   @Test
   public void testMultipleStagingTableCreation() throws Exception {
-    createCommonResources();
-
     String stagingTable1 = "staging_table_1";
     String stagingTable2 = "staging_table_2";
 
@@ -315,8 +306,6 @@ public class SdkTableCRUDTest extends BaseTableCRUDTest {
   /** Test that staging table creation fails when the schema doesn't exist. */
   @Test
   public void testStagingTableCreationWithNonExistentSchemaShouldFail() throws Exception {
-    createCommonResources();
-
     CreateStagingTable createStagingTableRequest =
         new CreateStagingTable()
             .catalogName(TestUtils.CATALOG_NAME)
